Modifies index route to return all possible move document types
