+++ pkg/models/move_documents_extractor_test.go
@@ -1,0 +1,33 @@
+package models_test
+
+import (
+	"github.com/transcom/mymove/pkg/models"
+	"github.com/transcom/mymove/pkg/testdatagen"
+
+	. "github.com/transcom/mymove/pkg/models"
+)
+
+func (suite *ModelSuite) TestFetchAllMoveDocumentsForMove() {
+	// When: there is a move and move document
+	move := testdatagen.MakeDefaultMove(suite.db)
+	sm := move.Orders.ServiceMember
+
+	assertions := testdatagen.Assertions{
+		MoveDocument: models.MoveDocument{
+			MoveID: move.ID,
+			Move:   move,
+		},
+		Document: models.Document{
+			ServiceMemberID: sm.ID,
+			ServiceMember:   sm,
+		},
+	}
+
+	testdatagen.MakeMoveDocument(suite.db, assertions)
+	testdatagen.MakeMovingExpenseDocument(suite.db, assertions)
+
+	docs, err := move.FetchAllMoveDocumentsForMove(suite.db)
+	if suite.NoError(err) {
+		suite.Len(docs, 2)
+	}
+}
